one commit

+++ src/main/java/com/example/spring/demo/service/UserService.java
@@ -9,9 +9,9 @@ public interface UserService {
 
     void saveUser(User user);
 
-    User getUser(int id);
+    User getUser(long id);
 
-    void deleteUser(int id);
+    void deleteUser(long id);
 
     void editUser(User user);
